refactor(desktop): keep memory routes out of oversized views
Move Memory destination ownership and Home stage padding policy into focused shared files so the UI polish stays within the product file line-count ratchet.

Verification: xcrun swift test --package-path desktop/macos/Desktop --filter 'DebouncedSearchCoordinatorTests|MemoryGraphRevisitTests|DesktopChatDriftGuardTests' (16 passed); xcrun swift build -c debug --package-path desktop/macos/Desktop (passed).
