+++ desktop/macos/Desktop/Sources/MainWindow/DesktopHomeView.swift
@@ -1229,32 +1229,6 @@ private struct HubSegmentedControl: View {
   }
 }
 
-enum MemoryHubDestination: Int, CaseIterable, Identifiable {
-  static let storageKey = "memoryHubDestination"
-
-  case memories
-  case conversations
-  case brainMap
-
-  var id: Int { rawValue }
-
-  var title: String {
-    switch self {
-    case .memories: return "Memories"
-    case .conversations: return "Conversations"
-    case .brainMap: return "Brain Map"
-    }
-  }
-
-  var icon: String {
-    switch self {
-    case .memories: return "brain.head.profile"
-    case .conversations: return "text.bubble"
-    case .brainMap: return "point.3.connected.trianglepath.dotted"
-    }
-  }
-}
-
 private struct MemoryHubPage: View {
   let appState: AppState
   let viewModelContainer: ViewModelContainer
